+++ framework/domain/src/main/java/is/codion/framework/domain/entity/DefaultStringFactory.java
@@ -47,6 +47,6 @@ final class DefaultStringFactory implements Function<Entity, String>, Serializab
 	}
 
 	private static String toString(Entity entity, Attribute<?> attribute) {
-		return attribute.name() + ": " + (entity.isNull(attribute) ? "null" : entity.string(attribute));
+		return attribute.name() + ": " + (entity.isNull(attribute) ? "null" : entity.format(attribute));
 	}
 }