+++ changelog.md
@@ -10,6 +10,7 @@ Codion Change Log
 ### is.codion.framework.domain
 - ConditionString.toString() renamed get().
 - AttributeDefinition.string(value) renamed format().
+- Entity.string(attribute) renamed format.
 ### is.codion.framework.db.core
 - EntityConnectionProvider.Builder, domainType.name() now used as clientType if none is specified.
 ### is.codion.swing.framework.model